# instagram-automation/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Instagram Automation API",
        "description": "AI-assisted Instagram content planning, approval, scheduling and publishing (official Meta APIs only).",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/instagram/status": {
            "get": {
                "tags": [
                    "instagram"
                ],
                "summary": "Instagram Status",
                "operationId": "instagram_status_api_instagram_status_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "include_limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Include Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstagramStatusOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/instagram/oauth/start": {
            "get": {
                "tags": [
                    "instagram"
                ],
                "summary": "Oauth Start",
                "description": "Returns the Instagram authorization URL. The browser navigates there.",
                "operationId": "oauth_start_api_instagram_oauth_start_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/instagram/test": {
            "post": {
                "tags": [
                    "instagram"
                ],
                "summary": "Test Connection",
                "operationId": "test_connection_api_instagram_test_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstagramStatusOut"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/instagram/refresh-token": {
            "post": {
                "tags": [
                    "instagram"
                ],
                "summary": "Refresh Token",
                "operationId": "refresh_token_api_instagram_refresh_token_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstagramStatusOut"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/instagram/disconnect": {
            "post": {
                "tags": [
                    "instagram"
                ],
                "summary": "Disconnect",
                "operationId": "disconnect_api_instagram_disconnect_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstagramStatusOut"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/sheets/status": {
            "get": {
                "tags": [
                    "google-sheets"
                ],
                "summary": "Sheets Status",
                "operationId": "sheets_status_api_sheets_status_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/sheets/sync": {
            "post": {
                "tags": [
                    "google-sheets"
                ],
                "summary": "Sheets Sync",
                "operationId": "sheets_sync_api_sheets_sync_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/sheets/preview": {
            "get": {
                "tags": [
                    "google-sheets"
                ],
                "summary": "Sheets Preview",
                "description": "Mock mode only: show the in-memory sheet so the sync can be inspected without Google.",
                "operationId": "sheets_preview_api_sheets_preview_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/dashboard": {
            "get": {
                "tags": [
                    "dashboard"
                ],
                "summary": "Dashboard",
                "operationId": "dashboard_api_dashboard_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "InstagramStatusOut": {
                "properties": {
                    "connected": {
                        "type": "boolean",
                        "title": "Connected"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "is_mock": {
                        "type": "boolean",
                        "title": "Is Mock"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "ig_user_id": {
                        "type": "string",
                        "title": "Ig User Id"
                    },
                    "account_type": {
                        "type": "string",
                        "title": "Account Type"
                    },
                    "login_type": {
                        "type": "string",
                        "title": "Login Type"
                    },
                    "token_source": {
                        "type": "string",
                        "title": "Token Source"
                    },
                    "token_expires_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Token Expires At"
                    },
                    "token_last_refreshed_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Token Last Refreshed At"
                    },
                    "scopes": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Scopes"
                    },
                    "last_success_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Success At"
                    },
                    "last_error": {
                        "type": "string",
                        "title": "Last Error"
                    },
                    "last_error_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Error At"
                    },
                    "oauth_available": {
                        "type": "boolean",
                        "title": "Oauth Available"
                    },
                    "publishing_limit": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Publishing Limit"
                    }
                },
                "type": "object",
                "required": [
                    "connected",
                    "status",
                    "is_mock",
                    "username",
                    "ig_user_id",
                    "account_type",
                    "login_type",
                    "token_source",
                    "token_expires_at",
                    "token_last_refreshed_at",
                    "scopes",
                    "last_success_at",
                    "last_error",
                    "last_error_at",
                    "oauth_available"
                ],
                "title": "InstagramStatusOut"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}